+++ .gitignore
@@ -26,7 +26,7 @@ yarn-error.log*
 
 # local env files
 .env*.local
-.env.d.ts
+env.d.ts
 
 
 # vercel
