--- locales/es.ts
@@ -1,52 +1,0 @@
-import { LenType } from './en'
-
-export const es: LenType = {
-  button: {
-    Button: 'Botón',
-    Link: 'Enlace que parece un botón',
-  },
-  errorsTr: {
-    errorNewPassword: {
-      refineMessage: 'Las contraseñas deben coincidir',
-    },
-    errorSignUp: {
-      isApprovedMassage: 'Por favor, lee y acepta los términos y condiciones',
-      refineMessage: 'Las contraseñas no coinciden',
-    },
-    errorValidationFields: {
-      maxChar: 'Número máximo de caracteres',
-      minChar: 'Número mínimo de caracteres',
-      passContain: 'La contraseña debe contener',
-      wrongUsername: 'Nombre de usuario incorrecto',
-    },
-  },
-  header: {
-    signIn: 'Iniciar sesión',
-    signUp: 'Registrarse',
-  },
-  language: 'Spanish',
-  signInLang: {
-    anAccount: '¿No tienes una cuenta?',
-    email: 'Correo electrónico',
-    forgoPassword: '¿Olvidaste la contraseña?',
-    password: 'Contraseña',
-    signIn: 'Iniciar sesión',
-    signUp: 'Registrarse',
-  },
-  signUpLang: {
-    IAgree: 'Estoy de acuerdo con los',
-    aler: 'Hemos enviado un enlace para confirmar tu correo electrónico a',
-    anAccount: '¿Tienes una cuenta?',
-    and: 'y',
-    email: 'Correo electrónico',
-    message: 'Por favor, lee y acepta los términos y condiciones',
-    password: 'Contraseña',
-    passwordConf: 'Confirmación de contraseña',
-    privPolicy: 'Política de privacidad',
-    signIn: 'Iniciar sesión',
-    signUp: 'Registrarse',
-    termServ: 'Términos de servicio',
-    username: 'Nombre de usuario',
-  },
-  test: 'Lorem Ipsum es simplemente texto de relleno de la impresión',
-}
