--- locales/ru.ts
@@ -1,52 +1,0 @@
-import { LenType } from './en'
-
-export const ru: LenType = {
-  button: {
-    Button: 'Кнопка',
-    Link: 'Ссылка, которая выглядит как кнопка',
-  },
-  errorsTr: {
-    errorNewPassword: {
-      refineMessage: 'Пароли должны совпадать',
-    },
-    errorSignUp: {
-      isApprovedMassage: 'Пожалуйста, прочитайте и примите условия',
-      refineMessage: 'Пароли не совпадают',
-    },
-    errorValidationFields: {
-      maxChar: 'Максимальное количество символов',
-      minChar: 'Минимальное количество символов',
-      passContain: 'Пароль должен содержать',
-      wrongUsername: 'Неправильное имя пользователя',
-    },
-  },
-  header: {
-    signIn: 'Войти',
-    signUp: 'Регистрация',
-  },
-  language: 'Russia',
-  signInLang: {
-    anAccount: 'Нет аккаунта?',
-    email: 'Электронная почта',
-    forgoPassword: 'Забыли пароль',
-    password: 'Пароль',
-    signIn: 'Войти',
-    signUp: 'Регистрация',
-  },
-  signUpLang: {
-    IAgree: 'Я согласен с',
-    aler: 'Мы отправили ссылку для подтверждения вашей электронной почты на',
-    anAccount: 'У вас есть аккаунт?',
-    and: 'и',
-    email: 'Электронная почта',
-    message: 'Пожалуйста, прочитайте и примите условия',
-    password: 'Пароль',
-    passwordConf: 'Подтверждение пароля',
-    privPolicy: 'Политикой конфиденциальности',
-    signIn: 'Войти',
-    signUp: 'Регистрация',
-    termServ: 'Условиями использования',
-    username: 'Имя пользователя',
-  },
-  test: 'Lorem Ipsum — это просто текст-«рыба» в полиграфии',
-}
