+++ public/locales/be.ts
@@ -1,0 +1,68 @@
+import {LenType} from './en'
+
+export const be: LenType = {
+    button: {
+        Button: 'Кнопка',
+        Link: 'Спасылка, якая выглядае як кнопка',
+    }, createNewPasswordForm: {
+        createNewPassword: 'Стварыць новы пароль',
+        newPassword: 'Новы пароль',
+        passwordConfirmation: 'Пацверджанне пароля',
+        passwordTerm: 'Ваш пароль павінен быць ад 6 да 20 знакаў',
+    },
+    errorsTr: {
+        errorEmail: {InvalidEmail: 'Няправільны электронны адрас'},
+        errorNewPassword: {
+            refineMessage: 'Паролі павінны супадаць',
+        },
+        errorSignUp: {
+            isApprovedMassage: 'Калі ласка, прачытайце і прыміце ўмовы',
+            refineMessage: 'Паролі не супадаюць',
+        },
+        errorValidationFields: {
+            maxChar: 'Максімальная колькасць знакаў',
+            minChar: 'Мінімальная колькасць знакаў',
+            passContain: 'Пароль павінен утрымліваць',
+            wrongUsername: 'Няправільнае імя карыстальніка',
+        },
+    },
+    forgoPasswordForm: {
+        again: ' Яшчэ Раз',
+        backSignIn: 'Вярнуцца да ўваходу',
+        email: 'Электронная пошта',
+        enterEmail: 'Увядзіце сваю электронную пошту, і мы вышлем вам дадатковыя інструкцыі',
+        forgotPassword: 'Забылі пароль',
+        linkSent: 'Спасылка была адпраўлена на электронную пошту.',
+        sendAgain: 'Калі вы не атрымалі электронны ліст, адпраўце спасылку яшчэ раз.',
+        sendLink: 'Адправіць Спасылку'
+    },
+    header: {
+        signIn: 'Увайсці',
+        signUp: 'Рэгістрацыя',
+    },
+    language: 'Belarus',
+    signInLang: {
+        anAccount: 'Няма ўліковага запісу?',
+        email: 'Электронная пошта',
+        forgoPassword: 'Забылі пароль',
+        password: 'Пароль',
+        signIn: 'Увайсці',
+        signUp: 'Рэгістрацыя',
+    },
+    signUpLang: {
+        IAgree: 'Я згодны з',
+        aler: 'Мы адправілі спасылку для пацверджання вашай электроннай пошты на',
+        anAccount: 'У вас ёсць уліковы запіс?',
+        and: 'і',
+        email: 'Электронная пошта',
+        message: 'Калі ласка, прачытайце і прыміце ўмовы',
+        password: 'Пароль',
+        passwordConf: 'Пацверджанне пароля',
+        privPolicy: 'Палітыкай прыватнасці',
+        signIn: 'Увайсці',
+        signUp: 'Рэгістрацыя',
+        termServ: 'Умовамі карыстання',
+        username: 'Імя карыстальніка',
+    },
+    test: 'Lorem Ipsum — гэта проста тэкст-запаўняльнік у друку',
+}
