+++ public/locales/ru.ts
@@ -1,0 +1,68 @@
+import {LenType} from './en'
+
+export const ru: LenType = {
+    button: {
+        Button: 'Кнопка',
+        Link: 'Ссылка, которая выглядит как кнопка',
+    },
+    createNewPasswordForm: {
+        createNewPassword: 'Создать новый пароль',
+        newPassword: 'Новый пароль',
+        passwordConfirmation: 'Подтверждение пароля',
+        passwordTerm: 'Ваш пароль должен быть от 6 до 20 символов',
+    },
+    errorsTr: {
+        errorEmail: {InvalidEmail: 'Неправильный адрес электронной почты'},
+        errorNewPassword: {
+            refineMessage: 'Пароли должны совпадать',
+        },
+        errorSignUp: {
+            isApprovedMassage: 'Пожалуйста, прочитайте и примите условия',
+            refineMessage: 'Пароли не совпадают',
+        },
+        errorValidationFields: {
+            maxChar: 'Максимальное количество символов',
+            minChar: 'Минимальное количество символов',
+            passContain: 'Пароль должен содержать',
+            wrongUsername: 'Неправильное имя пользователя',
+        },
+    }, forgoPasswordForm: {
+        again: ' Еще Раз',
+        backSignIn: 'Вернуться ко входу',
+        email: 'Электронная почта',
+        enterEmail: 'Введите свою электронную почту, и мы вышлем вам дальнейшие инструкции',
+        forgotPassword: 'Забыли пароль',
+        linkSent: 'Ссылка была отправлена по электронной почте.',
+        sendAgain: 'Если вы не получили письмо, отправьте ссылку еще раз.',
+        sendLink: 'Отправить Ссылку'
+    },
+    header: {
+        signIn: 'Войти',
+        signUp: 'Регистрация',
+    },
+    language: 'Russia',
+    signInLang: {
+        anAccount: 'Нет аккаунта?',
+        email: 'Электронная почта',
+        forgoPassword: 'Забыли пароль',
+        password: 'Пароль',
+        signIn: 'Войти',
+        signUp: 'Регистрация',
+    },
+    signUpLang: {
+        IAgree: 'Я согласен с',
+        aler: 'Мы отправили ссылку для подтверждения вашей электронной почты на',
+        anAccount: 'У вас есть аккаунт?',
+        and: 'и',
+        email: 'Электронная почта',
+        message: 'Пожалуйста, прочитайте и примите условия',
+        password: 'Пароль',
+        passwordConf: 'Подтверждение пароля',
+        privPolicy: 'Политикой конфиденциальности',
+        signIn: 'Войти',
+        signUp: 'Регистрация',
+        termServ: 'Условиями использования',
+        username: 'Имя пользователя',
+    },
+    test: 'Lorem Ipsum — это просто текст-«рыба» в полиграфии',
+}
