+++ public/locales/uk.ts
@@ -1,0 +1,67 @@
+import {LenType} from './en'
+
+export const uk: LenType = {
+    button: {
+        Button: 'Кнопка',
+        Link: 'Посилання, що виглядає як кнопка',
+    },
+    createNewPasswordForm: {
+        createNewPassword: 'Створити новий пароль',
+        newPassword: 'Новий пароль',
+        passwordConfirmation: 'Підтвердження пароля',
+        passwordTerm: 'Ваш пароль повинен бути від 6 до 20 символів',
+    },
+    errorsTr: {
+        errorEmail: {InvalidEmail: 'Неправильна електронна адреса'},
+        errorNewPassword: {
+            refineMessage: 'Паролі повинні збігатися',
+        },
+        errorSignUp: {
+            isApprovedMassage: 'Будь ласка, прочитайте і прийміть умови та положення',
+            refineMessage: 'Паролі не збігаються',
+        }, errorValidationFields: {
+            maxChar: 'Максимальна кількість символів',
+            minChar: 'Мінімальна кількість символів',
+            passContain: 'Пароль повинен містити',
+            wrongUsername: "Неправильне ім'я користувача",
+        }
+    }, forgoPasswordForm: {
+        again: ' Ще Раз',
+        backSignIn: 'Повернутися до входу',
+        email: 'Електронна пошта',
+        enterEmail: 'Введіть свою електронну пошту, і ми надішлемо вам подальші інструкції',
+        forgotPassword: 'Забули пароль',
+        linkSent: 'Посилання було надіслано на електронну пошту.',
+        sendAgain: 'Якщо ви не отримали листа, надіслати посилання ще раз',
+        sendLink: 'Надіслати Посилання'
+    },
+    header: {
+        signIn: 'Увійти',
+        signUp: 'Реєстрація',
+    },
+    language: 'Ukraine',
+    signInLang: {
+        anAccount: 'Немаєш акаунт?',
+        email: 'Електронна пошта',
+        forgoPassword: ' Забули пароль',
+        password: 'Пароль',
+        signIn: 'Увійти',
+        signUp: 'Реєстрація',
+    },
+    signUpLang: {
+        IAgree: 'Я погоджуюсь з',
+        aler: 'Ми відправили посилання для підтвердження вашої електронної пошти на',
+        anAccount: 'У вас є акаунт?',
+        and: 'та',
+        email: 'Електронна пошта',
+        message: 'Будь ласка, прочитайте і прийміть умови та положення',
+        password: 'Пароль',
+        passwordConf: 'Підтвердження пароля',
+        privPolicy: 'Політикою конфіденційності',
+        signIn: 'Увійти',
+        signUp: 'Реєстрація',
+        termServ: ' Умовами надання послуг',
+        username: "Ім'я користувача",
+    },
+    test: 'Lorem Ipsum — це просто текст-«риба» у поліграфії',
+}
