+++ src/features/auth/forgotPassword/ui/ForgotPasswordForm.tsx
@@ -3,12 +3,14 @@ import { useRef, useState } from 'react'
 import ReCAPTCHA from 'react-google-recaptcha'
 import { UseFormSetError } from 'react-hook-form'
 
+import { translate } from '@/shared/lib/langSwitcher'
 import { Nullable } from '@/shared/types'
 import { Button } from '@/shared/ui/button/button'
 import { Card } from '@/shared/ui/card'
 import { Input } from '@/shared/ui/input/input'
 import { clsx } from 'clsx'
 import Link from 'next/link'
+import { useRouter } from 'next/router'
 
 import s from './forgotPassword.module.scss'
 
@@ -43,8 +45,11 @@ export const ForgotPasswordForm = ({ onSubmit }: ForgotPasswordType) => {
     text: clsx(s.text),
     title: clsx(s.title),
   }
+  const { locale } = useRouter()
 
-  const { errors, handleSubmit, register, setError } = useForgotPassword()
+  const { errorsTr, forgoPasswordForm } = translate(locale)
+  const { errorEmail } = errorsTr
+  const { errors, handleSubmit, register, setError } = useForgotPassword(errorEmail)
 
   const siteKey = process.env.captchaSiteKey as string
 
@@ -61,37 +66,36 @@ export const ForgotPasswordForm = ({ onSubmit }: ForgotPasswordType) => {
 
   return (
     <Card className={classNames.card}>
-      <h1 className={classNames.title}>Forgot Password</h1>
+      <h1 className={classNames.title}>{forgoPasswordForm.forgotPassword}</h1>
       <form className={classNames.form} onSubmit={onSubmitHandler}>
         <Input
           className={classNames.input}
-          label={'Email'}
+          label={forgoPasswordForm.email}
           {...register('email')}
           error={errors.email?.message}
           placeholder={'Google@gmail.com'}
         />
-        <div className={classNames.text}>
-          Enter your email and we will send you further instruction
-        </div>
+        <div className={classNames.text}>{forgoPasswordForm.enterEmail}</div>
         {isSentLink && (
           <div className={classNames.hiddenText}>
-            The link has been sent by email.
+            {forgoPasswordForm.linkSent}
             <br />
-            If you do not receive an email send link again.
+            {forgoPasswordForm.sendAgain}
           </div>
         )}
         <Button className={classNames.button} fullWidth type={'submit'} variant={'primary'}>
-          {!isSentLink ? 'Send Link' : 'Send Link Again'}
+          {!isSentLink
+            ? forgoPasswordForm.sendLink
+            : forgoPasswordForm.sendLink + forgoPasswordForm.again}
         </Button>
       </form>
-
       <Button as={Link} className={classNames.btnLink} href={'/sign-in'} variant={'link'}>
-        Back to Sign In
+        {forgoPasswordForm.backSignIn}
       </Button>
 
       <div className={classNames.recaptcha}>
         <ReCAPTCHA
-          hl={'en'}
+          hl={locale}
           onChange={setToken}
           ref={captchaRef}
           sitekey={siteKey}
