+++ tsconfig.json
@@ -30,7 +30,7 @@
   },
   "include": [
     "next-env.d.ts",
-    ".env.d.ts",
+    "env.d.ts",
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts"
